Re-added stuff from the previous deleted commits.
Fancy window on login, disabled by default on the server_settings table. Need a .txt file located on gamedata folder (./habbopages/file.txt).

Tried to understand how the Sanction Thing works, and explained it on the source for someone wants to code this in the future.

Corrected the AvatarAspectUpdateMessageComposer packet with the right one.